tests: split core and psyneulink integration tests
