+++ dev_requirements.txt
@@ -1,4 +1,3 @@
-psyneulink >=0.9.1.0
 pydot
 pytest
 pytest-cov
