+++ psyneulink_requirements.txt
@@ -1,0 +1,1 @@
+psyneulink >=0.9.1.0
